+++ groups/templatetags/__init__.py
@@ -1,0 +1,1 @@
+"""Groups app template tags."""